requirements.txt
pandas
pytest
pillow
openpyxl
selenium
requests
beautifulsoup4
python-Levenshtein
